Refactor: import Img
